requirements.txt
numpy
matplotlib
opencv-python
scipy
scikit-image
tqdm
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python